Fix integer to long conversions errors on enumerations

+++ jocd/src/br/org/certi/jocd/core/Target.java
@@ -27,7 +27,7 @@ public class Target {
 
   public interface CoreRegister {
 
-    public int getValue();
+    public long getValue();
   }
 
   public enum State {
